moved run_local_experiment to within package in Experiment.py. Can now run experiment from within ipy without having to define any functions.
