+++ run_local_experiment.py
@@ -4,7 +4,7 @@ import sys
 import argparse
 from pathlib import Path
 
-import he6_cres_spec_sims.experiment as exp
+import he6_cres_spec_sims as spec_sims
 
 
 def main():
@@ -32,29 +32,9 @@ def main():
     )
 
     args = par.parse_args()
-    run_local_experiment(args.exp_dict_path)
+    spec_sims.run_local_experiment(args.exp_dict_path)
 
     return None
-
-
-def run_local_experiment(dict_path):
-
-    print(f"\n\n\n Beginning local simulation.\n\n\n")
-
-    experiment_name = Path(dict_path).stem
-
-    sim_experiment_params = json.load(open(dict_path))
-    sim_experiment_params["experiment_name"] = experiment_name
-
-    for key, val in sim_experiment_params.items():
-        print("{}: {}".format(key, val))
-
-    sim_experiment = exp.Experiment(sim_experiment_params)
-
-    print(f"\n\n\n Done running simulation. {sim_experiment_params}")
-
-    return None
-
 
 if __name__ == "__main__":
     main()
